requirements.txt
google-auth 
google-auth-httplib2 
google-auth-oauthlib 
google-api-python-client 
python-dotenv